+++ docs/misc/changelog.rst
@@ -8,7 +8,8 @@ For download links, please look at `Github release page <https://github.com/hill
 Pre-release 2.4.2a (WIP)
 ------------------------
 
-- added suport for Dict spaces in DummyVecEnv and SubprocVecEnv. (@AdamGleave) 
+- added support for Dict spaces in DummyVecEnv and SubprocVecEnv. (@AdamGleave) 
+- made SubprocVecEnv thread-safe by default; support arbitrary multiprocessing start methods. (@AdamGleave)
 
 Release 2.4.1 (2019-02-11)
 --------------------------